changed parseInt to parseDouble

+++ utils/src/main/java/com/epam/utils/StringUtils.java
@@ -2,6 +2,6 @@ package com.epam.utils;
 
 public class StringUtils {
     public static boolean isPositiveNumber(String str) {
-        return org.apache.commons.lang3.math.NumberUtils.isParsable(str) && Integer.parseInt(str) > 0;
+        return org.apache.commons.lang3.math.NumberUtils.isParsable(str) && Double.parseDouble(str) > 0;
     }
 }
